Refactor DistSQLStatementParserEngine.
