+++ shardingsphere-distsql/shardingsphere-distsql-parser/src/main/antlr4/imports/RULKeyword.g4
@@ -15,26 +15,26 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.distsql.parser.core.common;
+lexer grammar RULKeyword;
 
-import org.antlr.v4.runtime.CommonTokenStream;
-import org.antlr.v4.runtime.TokenStream;
-import org.apache.shardingsphere.distsql.parser.autogen.CommonDistSQLStatementParser;
-import org.apache.shardingsphere.sql.parser.api.parser.SQLParser;
-import org.apache.shardingsphere.sql.parser.api.visitor.ASTNode;
-import org.apache.shardingsphere.sql.parser.core.ParseASTNode;
+import Alphabet;
 
-/**
- * SQL parser for common dist SQL.
- */
-public final class CommonDistSQLParser extends CommonDistSQLStatementParser implements SQLParser {
-    
-    public CommonDistSQLParser(final TokenStream input) {
-        super(input);
-    }
-    
-    @Override
-    public ASTNode parse() {
-        return new ParseASTNode(execute(), (CommonTokenStream) getTokenStream());
-    }
-}
+WS
+    : [ \t\r\n] +
+    ;
+
+PREVIEW
+    : P R E V I E W
+    ;
+
+PARSE
+    : P A R S E
+    ;
+
+FORMAT
+    : F O R M A T
+    ;
+
+SQLString
+    : WS.*
+    ;
