+++ shardingsphere-distsql/shardingsphere-distsql-parser/src/main/antlr4/org/apache/shardingsphere/distsql/parser/autogen/KernelDistSQLStatement.g4
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-grammar CommonDistSQLStatement;
+grammar KernelDistSQLStatement;
 
 import Symbol, RALStatement, RDLStatement, RQLStatement;
