+++ shardingsphere-distsql/shardingsphere-distsql-parser/src/main/antlr4/org/apache/shardingsphere/distsql/parser/autogen/UtilityDistSQLStatement.g4
@@ -15,18 +15,13 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.distsql.parser.core.advanced;
+grammar UtilityDistSQLStatement;
 
-import org.antlr.v4.runtime.CharStream;
-import org.apache.shardingsphere.distsql.parser.autogen.AdvancedDistSQLStatementLexer;
-import org.apache.shardingsphere.sql.parser.api.parser.SQLLexer;
+import Symbol, RULStatement;
 
-/**
- * SQL lexer for advanced dist SQL.
- */
-public final class AdvancedDistSQLLexer extends AdvancedDistSQLStatementLexer implements SQLLexer {
-    
-    public AdvancedDistSQLLexer(final CharStream input) {
-        super(input);
-    }
-}
+execute
+    : (previewSQL 
+    | parseSQL
+    | formatSQL
+    ) SEMI?
+    ;
